- Resolve .bazelrc add/add conflict: keep our full Bazel config, incorporate
  CI output flags (--show_timestamps, --noshow_loading_progress, build:ci)
- Resolve MODULE.bazel conflict: keep our full Kotlin/Android Bazel setup,
  append rules_rust section from main (libsql JNI driver)
- Resolve .gitignore conflict: keep our version with Bazel entries, drop
  main's garbled *.tempuploads/ line
- Resolve GraphLoader.kt conflict: use main's compactFtsAfterBulkIndex() helper
- Resolve SqlDelightBlockRepository.kt conflict: drop duplicate FTS block
  (main already incorporated our ftsAutomergeDefault changes)
- Regenerate SQLDelight sources for all changes through v0.55.0

+++ kmp/src/commonMain/kotlin/dev/stapler/stelekit/repository/SqlDelightPropertyRepository.kt
@@ -144,7 +144,7 @@ class SqlDelightPropertyRepository(
             leftUuid = this.left_uuid,
             content = this.content,
             level = this.level.toInt(),
-            position = this.position.toInt(),
+            position = this.position,
             createdAt = Instant.fromEpochMilliseconds(this.created_at),
             updatedAt = Instant.fromEpochMilliseconds(this.updated_at),
             version = this.version,
